requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
firebase-admin
kaggle
google-cloud-storage
pydantic-settings
pydantic